refactor get_disks() functions
